+++ src/content/headsup/index.css
